awaiter compiles
